Split up authentication code into a separate server, away from the API code
