+++ backend/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "devStart": "nodemon server.ts"
+    "devStart": "concurrently --kill-others \"nodemon authserver.ts\" \"nodemon server.ts\""
   },
   "keywords": [],
   "author": "",
@@ -15,6 +15,7 @@
     "@types/jsonwebtoken": "^8.5.8",
     "@types/node": "^17.0.21",
     "bcrypt": "^5.0.1",
+    "concurrently": "^7.0.0",
     "dotenv": "^16.0.0",
     "express": "^4.17.2",
     "jsonwebtoken": "^8.5.1",
